导航栏分离化
