+++ app/app.vue
@@ -1,5 +1,12 @@
 <template>
-  <NuxtLayout>
-    <NuxtPage />
-  </NuxtLayout>
+  <div class="app-container">
+    <NavigationBar />
+    <NuxtLayout>
+      <NuxtPage />
+    </NuxtLayout>
+  </div>
 </template>
+
+<script setup>
+import NavigationBar from '~/components/navigation_bar.vue'
+</script>
