+++ app/pages/index.vue
@@ -1,43 +1,4 @@
 <template>
-    <!-- Navigation Bar -->
-    <nav class="navbar">
-      <div class="nav-container">
-        <div class="nav-logo">
-          <h1 class="nav-title">{{ getLocalizedText('transportationBureau') }}</h1>
-        </div>
-        <div class="nav-menu" :class="{ active: isMobileMenuOpen }">
-          <NuxtLink to="/" class="nav-link active">{{ getLocalizedText('home') }}</NuxtLink>
-          <NuxtLink to="/map" class="nav-link">{{ getLocalizedText('subwayLines') }}</NuxtLink>
-          <NuxtLink to="/station-info" class="nav-link">{{ getLocalizedText('stationInfo') }}</NuxtLink>
-          <NuxtLink to="/timetable" class="nav-link">{{ getLocalizedText('schedule') }}</NuxtLink>
-          <NuxtLink to="/copyright" class="nav-link">©</NuxtLink>
-        </div>
-        <div class="nav-right">
-          <div class="language-selector">
-            <select class="lang-dropdown" v-model="currentLanguage" @change="changeLanguage(currentLanguage)">
-              <option value="简体中文">
-                🇨🇳 简体中文
-              </option>
-              <option value="繁体中文">
-                🇭🇰 繁体中文
-              </option>
-              <option value="English">
-                English
-              </option>
-              <option value="日本語">
-                日本語
-              </option>
-            </select>
-          </div>
-          <div class="mobile-menu-toggle" @click="toggleMobileMenu">
-            <span class="hamburger"></span>
-            <span class="hamburger"></span>
-            <span class="hamburger"></span>
-          </div>
-        </div>
-      </div>
-    </nav>
-
     <!-- Top Card with Logo and City Name -->
     <div class="top-card">
       <div class="top-card-content">
@@ -70,26 +31,6 @@
 
 <script setup>
 import { ref } from 'vue'
-
-// Helper function to lighten a color
-const lightenColor = (color, percent) => {
-  // Parse the hex color
-  let R = parseInt(color.substring(1, 3), 16)
-  let G = parseInt(color.substring(3, 5), 16)
-  let B = parseInt(color.substring(5, 7), 16)
-
-  // Lighten the color
-  R = Math.min(255, R + (255 - R) * (percent / 100))
-  G = Math.min(255, G + (255 - G) * (percent / 100))
-  B = Math.min(255, B + (255 - B) * (percent / 100))
-
-  // Convert back to hex
-  const newR = Math.round(R).toString(16).padStart(2, '0')
-  const newG = Math.round(G).toString(16).padStart(2, '0')
-  const newB = Math.round(B).toString(16).padStart(2, '0')
-
-  return `#${newR}${newG}${newB}`
-}
 
 // Use useAsyncData with correct path handling for server-side
 const { data: configData } = await useAsyncData('config', async () => {
@@ -125,8 +66,6 @@ const currentLanguage = ref('简体中文')
 
 const changeLanguage = (langCode) => {
   currentLanguage.value = langCode
-  // Here you could add logic to change the language of the entire page
-  // For now, we're just updating the current language value
 }
 
 // Function to get localized text based on current language
@@ -137,36 +76,6 @@ const getLocalizedText = (key) => {
       '繁体中文': `${cityName}交通局`,
       'English': `${cityName} Transportation Bureau`,
       '日本語': `${cityName}交通局`
-    },
-    'home': {
-      '简体中文': '首页',
-      '繁体中文': '首頁',
-      'English': 'Home',
-      '日本語': 'ホーム'
-    },
-    'subwayLines': {
-      '简体中文': '地铁线路',
-      '繁体中文': '地鐵路線',
-      'English': 'Subway Lines',
-      '日本語': '地下鉄路線'
-    },
-    'stationInfo': {
-      '简体中文': '站点信息',
-      '繁体中文': '站點資訊',
-      'English': 'Station Info',
-      '日本語': '駅情報'
-    },
-    'schedule': {
-      '简体中文': '时刻表',
-      '繁体中文': '時刻表',
-      'English': 'Schedule',
-      '日本語': '時刻表'
-    },
-    'ticketInfo': {
-      '简体中文': '票务信息',
-      '繁体中文': '票務資訊',
-      'English': 'Ticket Info',
-      '日本語': '切符情報'
     },
     'welcome': {
       '简体中文': '欢迎访问',
@@ -200,10 +109,6 @@ const getLocalizedText = (key) => {
       ? localizedText[key]['简体中文'] 
       : key
 }
-
-const toggleMobileMenu = () => {
-  isMobileMenuOpen.value = !isMobileMenuOpen.value
-}
 </script>
 
 <style>
@@ -219,160 +124,7 @@ body {
   background-color: #f5f5f5;
 }
 
-/* Navigation Bar styles */
-.navbar {
-  background-color: var(--primary-color, #0047AB); /* Dynamic primary color from config with fallback */
-  color: white;
-  position: sticky;
-  top: 0;
-  z-index: 100;
-  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-}
 
-.nav-container {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 15px 20px;
-  max-width: 1200px;
-  margin: 0 auto;
-}
-
-.nav-logo {
-  display: flex;
-  align-items: center;
-}
-
-.nav-title {
-  font-size: 18px;
-  font-weight: bold;
-  margin: 0;
-  color: white;
-}
-
-.nav-menu {
-  display: flex;
-  gap: 25px;
-  align-items: center;
-}
-
-.nav-link {
-  color: white;
-  text-decoration: none;
-  font-size: 14px;
-  padding: 8px 12px;
-  border-radius: 4px;
-  transition: background-color 0.3s;
-}
-
-.nav-link:hover {
-  background-color: rgba(255, 255, 255, 0.1);
-}
-
-.nav-right {
-  display: flex;
-  align-items: center;
-  gap: 15px;
-}
-
-
-
-.language-selector {
-  display: flex;
-  align-items: center;
-}
-
-.lang-dropdown {
-  background-color: rgba(255, 255, 255, 0.15);
-  color: white;
-  border: 1px solid rgba(255, 255, 255, 0.3);
-  padding: 5px 10px;
-  font-size: 12px;
-  border-radius: 12px;
-  cursor: pointer;
-  outline: none;
-  appearance: none; /* Remove default dropdown arrow */
-  -webkit-appearance: none;
-  -moz-appearance: none;
-  background-image: url("data:image/svg+xml;charset=UTF-8,%3csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 24 24' fill='white'%3e%3cpath d='M7 10l5 5 5-5z'/%3e%3c/svg%3e");
-  background-repeat: no-repeat;
-  background-position: right 8px center;
-  background-size: 12px;
-  padding-right: 25px;
-  min-width: 100px;
-}
-
-/* Ensure text color in dropdown options */
-.lang-dropdown option {
-  color: #333; /* Dark color for options to ensure readability */
-  background-color: white; /* White background for options */
-}
-
-.lang-dropdown:hover {
-  background-color: rgba(255, 255, 255, 0.25);
-}
-
-.lang-dropdown:focus {
-  outline: none;
-}
-
-/* Mobile menu toggle */
-.mobile-menu-toggle {
-  display: none;
-  flex-direction: column;
-  cursor: pointer;
-  padding: 5px;
-}
-
-.hamburger {
-  width: 25px;
-  height: 3px;
-  background-color: white;
-  margin: 3px 0;
-  transition: 0.3s;
-}
-
-/* Tablet and Mobile responsiveness */
-@media (max-width: 768px) {
-  .nav-container {
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 15px;
-  }
-  
-  .nav-menu {
-    display: none;
-    width: 100%;
-    flex-direction: column;
-    background-color: var(--primary-color, #0047AB);
-    padding: 10px 0;
-    border-radius: 4px;
-  }
-  
-  .nav-menu.active {
-    display: flex;
-  }
-  
-  .nav-link {
-    padding: 10px 20px;
-    width: 100%;
-  }
-  
-  .nav-right {
-    width: 100%;
-    justify-content: space-between;
-  }
-  
-
-  
-  .mobile-menu-toggle {
-    display: flex;
-  }
-  
-  .logo-section {
-    padding-top: 20px;
-  }
-}
 
 /* Top Card styles */
 .top-card {
